Fix inconsistency with format specification for IDs
IDs starting with 0 are considered invalid although, they are allowed by the specification. The standard clearly specifies that IDs can start with 0:  "To accommodate the use of empty nodes for the analysis of ellipsis in the enhanced dependency representation, we adopt a further extension of the indexing scheme from v2. It is possible to insert one or more empty nodes indexed i.1, i.2, etc. immediately after a word with index i (where i = 0 for sentence-initial empty nodes)." taken from http://universaldependencies.org/format.html#words-tokens-and-empty-nodes